model TreeItem to model so imports do not collide
